Criar robôs como entidades externas e implementar notificações (#9)
* criar notificacoes

* adicionar direcionalidade nas notificacoes

* acabar sistema de escalonamento, falta remover prints

* remover ficheiro de backup da bd

* remover alteração redundante rovo

+++ Terceira Fase/ArmazemInteligente/src/main/java/Database/RotaDAO.java
@@ -67,6 +67,24 @@ public class RotaDAO {
         return  idsRobosEmTransito;
     }
 
+    public boolean roboEmTransito(int idRobo){
+        Connection conn = ConnectionPool.getConnection();
+
+        try (Statement sta = conn.createStatement()) {
+            String sql = "SELECT * from Rota where idRobo="+idRobo+";";
+            ResultSet rs = sta.executeQuery(sql);
+            if(rs.next()) {
+                return true;
+            }
+        } catch (SQLException e) {
+            e.printStackTrace();
+        }finally {
+            ConnectionPool.releaseConnection(conn);
+        }
+
+        return false;
+    }
+
     public Coordenadas getProximoPasso(int idRobo){
         Connection conn = ConnectionPool.getConnection();
         Coordenadas proximoPasso = null;
@@ -74,17 +92,18 @@ public class RotaDAO {
         try (Statement stm = conn.createStatement()) {
             String sql = "SELECT * from Rota where idRobo="+idRobo+";";
             ResultSet rs = stm.executeQuery(sql);
-            rs.next();
-            String rotaCodificada = rs.getString("valor");
-            List<Coordenadas> rotaDescodificada = descodificaRota(rotaCodificada);
-            proximoPasso = rotaDescodificada.remove(0);
+            if(rs.next()) {
+                String rotaCodificada = rs.getString("valor");
+                List<Coordenadas> rotaDescodificada = descodificaRota(rotaCodificada);
+                proximoPasso = rotaDescodificada.remove(0);
 
-            // se a lista estiver vazia depois de remover o próximo passo, remove-se a entrada na tabela
-            if(rotaDescodificada.isEmpty()){
-                String queryDeEliminacao = "DELETE from Rota where idRobo="+idRobo +";";
-                stm.executeUpdate(queryDeEliminacao);
-            }else{ // atualiza a rota
-                adicionaRota(idRobo,rotaDescodificada);
+                // se a lista estiver vazia depois de remover o próximo passo, remove-se a entrada na tabela
+                if (rotaDescodificada.isEmpty()) {
+                    String queryDeEliminacao = "DELETE from Rota where idRobo=" + idRobo + ";";
+                    stm.executeUpdate(queryDeEliminacao);
+                } else { // atualiza a rota
+                    adicionaRota(idRobo, rotaDescodificada);
+                }
             }
         } catch (SQLException e) {
             e.printStackTrace();
